requirements.txt
discord.py>=2.3
anthropic>=0.40
reportlab>=4.0
python-dotenv>=1.0

# dev / test
pytest>=8.0
pytest-asyncio>=0.23
pytest-mock>=3.12
